pyproject.toml
[project]
name = "rag-lab-de24-john-sandsjo"
version = "0.1.0"
description = "Individual project for course; Dataplattformar, maskininlärning och artificiell intelligens"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.2",
    "fastapi>=0.123.7",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.26.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.0",
]
